Use long long for 64-bit int on Windows

+++ src/repast4py/borders.h
@@ -40,7 +40,7 @@ struct Offset {
 
 template<>
 struct Offset<R4Py_DiscretePoint> {
-    constexpr static long value = 1;
+    constexpr static long long value = 1;
 };
 
 template<>
