+++ src/repast4py/geometry.cpp
@@ -16,7 +16,7 @@ R4Py_DiscretePoint* create_point(PyTypeObject* pt_type, const Point<R4Py_Discret
 
 bool point_equals(R4Py_DiscretePoint* pt, const Point<R4Py_DiscretePoint>& coords) {
     if (pt) {
-        long* data = (long*)PyArray_DATA(pt->coords);
+        long long* data = (long long*)PyArray_DATA(pt->coords);
         //printf("%lu,%lu,%lu  -- %lu,%lu,%lu\n", data[0], data[1], data[2],
         //    coords.x, coords.y, coords.z);
         return data[0] == coords.x && data[1] == coords.y && data[2] == coords.z;
@@ -25,14 +25,14 @@ bool point_equals(R4Py_DiscretePoint* pt, const Point<R4Py_DiscretePoint>& coord
 }
 
 void extract_coords(R4Py_DiscretePoint* pt, Point<R4Py_DiscretePoint>& coords) {
-    long* data = (long*)PyArray_DATA(pt->coords);
+    long long* data = (long long*)PyArray_DATA(pt->coords);
     coords.x = data[0];
     coords.y = data[1];
     coords.z = data[2];
 }
 
 void update_point(R4Py_DiscretePoint* pt, const Point<R4Py_DiscretePoint>& coords) {
-    long* data = (long*)PyArray_DATA(pt->coords);
+    long long* data = (long long*)PyArray_DATA(pt->coords);
     data[0] = coords.x;
     data[1] = coords.y;
     data[2] = coords.z;
